requirements.txt
Pillow==8.4.0
matplotlib==3.4.3
numpy==1.21.4
scipy==1.7.3